backend/elastic_seacher/requirements.txt
fastapi==0.90.0
starlette==0.23.0
uvicorn==0.22.0
elasticsearch==7.17.3
python-dotenv==0.21.0
loguru==0.6.0
sentence-transformers==2.2.2
sentencepiece==0.2.0
requests==2.28.2
pydantic==1.10.4
numpy==1.21.6
pandas==1.1.5
